backend/requirements.txt
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
